move more elasticsearch logic into elasticsearch module
