+++ packages/api/index.js
@@ -45,6 +45,6 @@ module.exports.handler = (event, context, cb) => {
   var page = query.page || 1
   var limit = query.limit || 100
 
-  satlib.api(event.path, endpoint, query, page, limit, respond)
+  satlib.api(event.path, endpoint, query, satlib.es, page, limit, respond)
   
 }
